requirements.txt
mysql-connector-python==8.0.20
psycopg2-binary==2.8.5
pandas==1.0.1